Refactor fetchallergens
Refactor the fetch method for allergens, to 1. include id, and 2. to only add new allergens to the list, meaning we don't readd the same object multiple times, since a new object is made each time we call fetch, hence why we compare on name. Maybe not be optimal computation wise, but it is a solution that works