Add camera-feature to main
Adding the camera-feature to main. We have made a repository for images that can used for general purposes when one wants to upload an image. We have integrated the camera button in the UI. There is a button in the post_dish_page for the camera. There have also been created a camera_state_controller to take pictures and a camera_widget that contains the buttons, etc. We make one big juicy commit due to problems with netlify, which made our initial branch cursed and forced us to make a new branch, where we had to copy in all our code. The code has also been tested in camera_widget_test.dart.

Co-authored-by: Markus <mbrh@itu.dk>